refactored forgotpassword getmapping
