+++ src/main/resources/templates/users/login.html
@@ -26,8 +26,7 @@
     <br>
     <button class="btn btn btn-primary" type="submit">Log In</button>
     <p class="message">Not registered? <a th:href="@{/register}">Create an account</a></p>
-    <br>
-    <p class="forgot-password"><a th:href="@{/forgot-password}">Forgot your password?</a></p>
+    <p class="message"><a th:href="@{/forgot-password}">Forgot Password</a></p>
 </form>
 
 </div>
